create a database
I'm just throwing everything into the database and I'll handle it there.
It's certainly not the best solution but it maintains a perfect record
in the sense that, if anybody wants to reimplement the data parsing code
they'll have the raw data to sample.  And there's no chance that I won't
include some data that I think it's not important right now but may turn
up important later.