add some scripts to work with old CAN data
We have these CSV files from the previous competitions that we can use
to test our current telemetry setup.

The Python script can translate the data to the asc and log formats that
`can-utils` support to replay them as if the data is coming from the
CANbus directly.

The SQL can be used to read the data directly into PostgreSQL and have
it be translated again to raw CANbus data.  This is much faster but does
not test the whole stack.